fix: address CodeQL-flagged workflow permissions and a real 32-bit multiplication overflow
- build.yml/nexus-upload.yml: add explicit least-privilege `permissions:
  contents: read` (CodeQL alerts #1/#2 -- neither workflow writes back to
  the repo or any GitHub API resource).
- SoundBank.cpp: ReadHeader()'s dependencyCount * dependencySize bounds
  check multiplied two raw `unsigned int` header fields directly (unlike
  the sibling entryOffset/checksumOffset checks two lines above, which
  multiply against `sizeof(...)` and promote to size_t automatically) --
  a crafted sound bank could overflow the 32-bit product back under the
  512-byte bound and pass the check despite the real value being far
  larger (CodeQL alert #8). Cast to size_t before multiplying.

CodeQL alerts #3-#7 (lzx.c narrow/wide loop-condition comparisons) were
dismissed as won't-fix -- vendored third-party OpenAssetTools upstream
decompression code, not maintained by this fork.

+++ tools/iw5oat/src/ObjLoading/ObjContainer/SoundBank/SoundBank.cpp
@@ -172,7 +172,7 @@ bool SoundBank::ReadHeader()
         return false;
     }
 
-    if (m_header.dependencyCount * m_header.dependencySize > sizeof(SoundAssetBankHeader::dependencies))
+    if (static_cast<size_t>(m_header.dependencyCount) * m_header.dependencySize > sizeof(SoundAssetBankHeader::dependencies))
     {
         con::error("Invalid sndbank dependency sizes (count is {}; size is {})", m_header.dependencyCount, m_header.dependencySize);
         return false;
